% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quran_config.R
\name{print.QuranConfig}
\alias{print.QuranConfig}
\title{Print a QuranConfig object}
\usage{
\method{print}{QuranConfig}(x, ...)
}
\arguments{
\item{x}{A \code{QuranConfig} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print a QuranConfig object
}
