% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_analytic_config.R
\name{print.TransAnalyticConfig}
\alias{print.TransAnalyticConfig}
\title{Print a TransAnalyticConfig object}
\usage{
\method{print}{TransAnalyticConfig}(x, ...)
}
\arguments{
\item{x}{A \code{TransAnalyticConfig} object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print a TransAnalyticConfig object
}
