% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trans_en_pickthall}
\alias{trans_en_pickthall}
\title{Quran Translation: Mohammed Marmaduke Pickthall (English)}
\format{
A \code{translationList} (S3) object. See \code{\link{trans_en_sahih}}
  for full format description.
}
\source{
\url{https://tanzil.net/trans/en.pickthall}
}
\usage{
trans_en_pickthall
}
\description{
English translation of the Quran by Mohammed Marmaduke Pickthall, downloaded
from \url{https://tanzil.net/trans/en.pickthall}. A \code{translationList}
object produced by \code{\link{tanzil_translation}}.
}
\keyword{datasets}
