% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trans_en_yusufali}
\alias{trans_en_yusufali}
\title{Quran Translation: Abdullah Yusuf Ali (English)}
\format{
A \code{translationList} (S3) object. See \code{\link{trans_en_sahih}}
  for full format description.
}
\source{
\url{https://tanzil.net/trans/en.yusufali}
}
\usage{
trans_en_yusufali
}
\description{
English translation of the Quran by Abdullah Yusuf Ali, downloaded from
\url{https://tanzil.net/trans/en.yusufali}. A \code{translationList} object
produced by \code{\link{tanzil_translation}}.
}
\keyword{datasets}
