% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trans_id_indonesian}
\alias{trans_id_indonesian}
\title{Quran Translation: Indonesian Ministry of Religious Affairs (Indonesian)}
\format{
A \code{translationList} (S3) object. See \code{\link{trans_en_sahih}}
  for full format description.
}
\source{
\url{https://tanzil.net/trans/id.indonesian}
}
\usage{
trans_id_indonesian
}
\description{
Indonesian translation of the Quran by the Indonesian Ministry of Religious
Affairs, downloaded from \url{https://tanzil.net/trans/id.indonesian}. A
\code{translationList} object produced by \code{\link{tanzil_translation}}.
}
\keyword{datasets}
