% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trans_ms_basmeih}
\alias{trans_ms_basmeih}
\title{Quran Translation: Abdullah Muhammad Basmeih (Malay)}
\format{
A \code{translationList} (S3) object. See \code{\link{trans_en_sahih}}
  for full format description.
}
\source{
\url{https://tanzil.net/trans/ms.basmeih}
}
\usage{
trans_ms_basmeih
}
\description{
Malay translation of the Quran by Abdullah Muhammad Basmeih, downloaded from
\url{https://tanzil.net/trans/ms.basmeih}. A \code{translationList} object
produced by \code{\link{tanzil_translation}}.
}
\keyword{datasets}
