% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreQuran-package.R
\docType{package}
\name{XploreQuran-package}
\alias{XploreQuran}
\alias{XploreQuran-package}
\title{XploreQuran: Quran Translation Analytic}
\description{
Text mining and analytic for Quran translation from tanzil.net with a Shiny web application.
}
\keyword{internal}
