% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_surah.r
\name{tf_trans}
\alias{tf_trans}
\title{Create term frequency by surah}
\usage{
tf_trans(tanzil_trans_object, surah_number = (1:114))
}
\arguments{
\item{tanzil_trans_object}{Tanzil translation object created with `tanzil_translation()` function}

\item{surah_number}{Surah number. Default to all 1-114 surah. Can refer to `quran_index` for surah number and Quran index.}
}
\value{
A term frequency data frame for selected surah
}
\description{
Create term frequency by surah
}
\examples{
# Example using package default data (if data is built)
# data(trans_en_sahih, package = "XploreQuran")
# tf_trans(trans_en_sahih, surah_number = 1)


}
